// src/app/shared/model/user.model.ts
export class UserModel {
  'mail': string;
  'password'?: string;
  'stayConnected'?: boolean;
  'session_id'?: string;
  'name'?: string;
  'firstname'?: string;
  'status'?: string;
  'school_company'?: string;
  'newsletter'?: boolean;
  'verified'?:boolean;
  'connection_type'?: string;
  'langage'?: string;
  'iat'?: number;
  'exp'?: number;
}